Firmware and rHTC development.
Started rewriting firmware.
Moved to ATmega324P.
Updated V-USB.
Moved LCD code to native avr-libc delay functions.
Implemented USB interface for rHTC.
Minor improvements in rHTC.

+++ firmware/temperature.h
@@ -1,3 +1,4 @@
+
 #define MAXTS 3
-uint8_t gtemp (int);
+uint16_t gtemp (int);
 uint8_t search_sensors (void);
